feat: use hatch as build system
Hatch is a convenient Python project manager. With this updated
pyproject.toml, pip is able to install the project with custom scripts
as before.

The big benefits come after that, Hatch runs commands in dedicated
virtual environments, providing some by default for tests and
formatting.

Running
    `ruff check . --statistics`
becomes
    `hatch fmt --check --linter --statistics`

And the command is automatically executed in a dedicated venv. It seems
the default linter config for hatch doesn't match what's currently used,
its founding 325 errors :|