feat: copy beancount v2.0 regression test helper
With the switch to v3, regression_pytest.py went MIA and completely
broke Reds' importers test suite.

Tests will probably need to be updated to better match beangulp and
beancount new flow, but in the meantime it would be better to have tests
that actually work.

In preparation for that, add this file as is to the project:
https://raw.githubusercontent.com/beancount/beancount/refs/heads/v2/beancount/ingest/regression_pytest.py